Overhaul of devices and filesystems
There is now a Device class, which all devices will extend. I also implemented BlockDevice and PIOBlockDevice. This way, reading and writing to disks is abstracted away so that a proper PCI disk driver can be written in the future without having to change everything else.

I did the same thing with the filesystem, with a base Filesystem class and an Ext2Filesystem class. This way, virtual filesystems can be easily implemented in the future.

Because of this, though, I just commented out everything in the shell commands, because I didn't feel like refactoring everything in there. So, I don't really know if reading files works properly yet. I should probably just completely rewrite the ext2 driver.