Completely rewritten memory management (see details)
Memory management now uses a linked list of MemoryRegion objects to keep track of regions of memory (which could easily be changed to a binary tree or something more efficient later, since it's abstracted away).

Regions have a start address and size, along with some other properties like their related region (the physical region corresponding to a virtual region / vice versa), whether or not they're marked CoW, if they're marked reserved according to the system memory map, and a couple other properties used for semantics.

This has the result of making memory management much easier to work with. For instance, the CoW code is now leagues simpler than it was before; the memory regions in the parent/child are simply marked CoW / read only and then when they need are all copied or destroyed, the old physical region is freed.

This new memory system also takes the memory map provided by GRUB into account, so accidentally using reserved / broken memory shouldn't happen, and the OS knows how much memory it has available to it now.