Kernel: Initial separation of Processes and Threads
Processes and Threads are now separate, so that hopefully, in the near future, processes can have multiple threads.

The scheduler now keeps track of processes in a vector, and instead of a round-robin scheduling approach, threads are inserted into a queue when needed.

Each process still has its own memory space, file descriptors, etc; the only thing that threads maintain is their own userspace and kernel stacks, and the main thread of a process will handle incoming signals.

Additionally, shared pointers are now used to refer to processes and threads throughout the kernel instead of raw pointers.