Kernel: Fix MemoryMap region allocation
Since MemoryMap would allocate a new MemoryRegion in the heap while allocating a new region, this would sometimes cause corruption of the MemoryMap. This was because allocating a new MemoryRegion may require allocating more pages for the heap, which would in turn allocate another region in the MemoryMap. This left the first call to allocate_region having an incorrect understanding of the map. This was fixed by just allocating the new region before iterating through the map.