Quack/libsound/libduck: Change how audio server works
Before, we were sending an IPC request to the audio server (Quack) every time we want to queue samples from a shared buffer (which is pretty time consuming & resulted in a lot of memory churn).

Now, we have a new structure called AtomicCircularQueue which is set up in a memory region shared between the audio server and client. This queue can be operated on atomically, allowing us to eliminate the need to send an IPC request every time we want to queue samples; now we can just simply push samples into the buffer and the audio server will pop them as it needs them.

This results in a way more efficient audio server. I still need to work to bring latency down and fix resampling, but it's progress.