Kernel: Fix thread memory leak
Threads were not being cleaned up properly most of the time due to ARC pointers not being cleaned up properly in a couple of places.

First, the old_thread pointer in TaskManager was not being reset before preemption, so if the thread was never going to be switched back to, that reference was leaked.

Second, when a thread exits, it could have a smart pointer to itself on the stack. But since it was dying, the stack is never cleaned up properly, and the reference was leaked.