Kernel: [WIP] Use a tree for VMSpace instead of a linked list
This updates VMSpace to use a tree structure to store regions instead of a linked list, which simplifies things and (theoretically) speeds them up as well.

In practice, however, this seems to make things *slower* when compiling without optimizations - probably because the VMSpace trees are incredibly unbalanced due to the fact that new regions are pretty much being inserted sequentially.

Theoretically, once kstd::map is self-balancing, it should speed things up quite a bit. Then, I'll merge this into the main branch.