Kernel: Accurate timekeeping with TSC
Before, the RTC or PIT interrupt was used to keep time. However, this could be fairly inaccurate if the timekeeper had to wait for other interrupts or anything else.

By measuring the speed at which the TSC increments at boot time by using the PIT, we can use it instead as a reliable way to track the elapsed time since boot.